Updated openstack/openstack Project: openstack/openstack-manuals  89fcd8106abe043dab6d4db79d8dd9796b07563f
Refactoring HP 3PAR Cinder Driver to be HPE

The HP 3PAR Cinder driver was recently refactored to be HPE.

The same refactoring is being done to the documentation with
this patch. It is being done because the company managing the
HP 3PAR Cinder driver is now Hewlett Packard Enterprise (HPE).

Closes-Bug: #1516076
Change-Id: Ie3857fdeb7aee371a3f081d93c423dcfff764cd2