Updated openstack/openstack Project: openstack/neutron-lbaas  ae573d11bcb27d8ae17ddb25794d2d3e5fdd4438
Disable automatic config file discovery for pep8 target

When trying to run pep8 target on a system with installed
/etc/neutron/neutron.conf file, neutron-db-manage may fail due to
access denied to the configuration file. This is because oslo.config
attempts to discover configuration files automatically if they are
not passed with --config-file argument.

To disable auto discovery, we pass --config-file to the tool.

Change-Id: I89b5cbae1fcbf662cba6bc724681d0cfeeb0eb9c