Updated openstack/openstack Project: openstack-infra/shade  ce5d6ea29681556c11a6199c58599fd1c9a47661
Allow specifying cloud name to ansible tests

It would be *really* cool and useful and stuff to be able to run
the Ansible playbooks against more than just devstack. This change
allows that. Now you can send the -c option to the tox command line:

    tox -e ansible -- -c coolcloud

Or, specific tests:

    tox -e ansible -- -c coolcloud auth network

Going against devstack is still:

    tox -e ansible
    tox -e ansible auth network

Change-Id: I666a09aee3d283865c3813d67edfb75123057c22