Updated openstack/openstack Project: openstack/ironic  e5a959724925c8359be9fcdfde6ccc61fe30c38a
Add manual cleaning to state machine

This adds transitions to the state machine so that manual cleaning can
be performed on a node in the manageable provision state via the
'clean' verb. After cleaning is successfully finished, the node will
be in the manageable provision state again (via the 'manage' verb).

The state machine diagram was updated to reflect this.

Change-Id: Ifacdcceb65d1384351aa832042a7471bb48f0894
Implements: blueprint manual-cleaning