Updated openstack/openstack Project: openstack-dev/devstack  85ce0d2711ad34ea6250d93f5da4f991209b6a7a
remove precise from the supported list

We haven't been testing master on precise for a long time, and changes
are coming that won't work on precise. If people want to keep running
on precise they should realize they are on their own with it. And that
we won't block any changes that use it.

Change-Id: I3697f1c2409ad08f49793dabb37011606188e6f6