Updated openstack/openstack Project: openstack/api-site  e1db24865ed4d7056a27d2d95dd3e6171904b1ac
Reverse order of values in for compute boot_index

The current docs for the boot_index parameter to block_device_mapping_v2
are backwards, suggesting that -1 be given for boot volume and 0 for
other volumes. The opposite is true according to both the novaclient
shell code and doc/source/block_device_mapping.rst in the nova source
tree. The new paragraph is taken directly from that nova source
document.

Change-Id: Id050903a8c71cac6324007429dfd30f40da10c48