Updated openstack/openstack Project: openstack/cinder  acf881e0aa6b3f455cb7d6aa77b9d31d9e5d4eb9
Fix swap_volume for case without migration

Fix swap_volume so that even when a volume-migration does not
occur, the volumes don't get stuck in "attaching" and "detaching".
Previously the swap_volume was changed in
https://review.openstack.org/#/c/101932 and
https://review.openstack.org/#/c/101933 to move the final
attach and detach logic to Cinder, however those changes only fix
the migration path.  This change calls Cinder's "attach" and "detach"
for the non-migration path to update the volumes' states.

Change-Id: I242c2c7c4a7197bbce04d0b3d75688f989ea1fd5
Closes-Bug: #1489744