Updated openstack/openstack Project: openstack/cinder  c8756996d19c58c536032ef884028164a8d9c5ef
Deprecate *_multipath_enabled flag for IBM drivers

Deprecate the *_multipath_enabled from the FlashSystem and
Storwize/SVC driver. When enabled, this flag was intended to create
a single mapping between the compute node and the storage device for
FC connections. However, all available storage WWNNs were still being
mapped to the compute node, and instead enabling this flag was
causing a terminate_connection to not cleanup all LUN mappings
(because Nova did not know of all the target_wwns that were actually
mapped). This can be seen by enabling the flag, connecting a volume,
disconnecting the volume and observing the stray LUN mappings under
'/dev/disk/by-path'.

Currently we do not have a way to limit the storage WWNNs which are
mapped during a FC connection, so we will ignore the config to fix
this broken behavior, and then remove the config next release.

DocImpact

Change-Id: I0019c1e0c0dbb0cd48a7f389376480e7fa1b1e07
Closes-Bug: #1516778