Updated openstack/openstack Project: openstack/ironic  e3a2e7f1b63bcc57c48bf2b60bee3e1cea79548c
Adds test case for the iscsi_ilo recreate boot iso

When using web server for hosting images for ilo driver,
take_over should recreate the boot ISO, if it was created
by Ironic. This fix is done as part of the feature patch
I5481e705892f66cb3184ac16ff56c123f5c69087 in boot.py from
lines 117-122.
The test case for this bug is missing in the feature patch
and hence adding it here in this patch.

Related-Bug: #1510425

Change-Id: I6b7f3901f773a5aba906c50dcbd574412f7b6af3