Updated openstack/openstack Project: openstack/nova  ea725fd6ba0c2a4236f172dced1eca2e9a3df2cc
Change warn to debug logs when migration context is missing

MigrationContext record missing is currently a regular situation. There
will be no migration record at all (and thus no MigrationContext record)
in case of a rebuild on the same host, and in case of evacuate request,
which explicitly specifies the target host.

With this in mind we demote the log messages that point out the missing
migration context to debug.

Change-Id: I96b6c859a9d450707f997e43e37f770e2a9dea7a