Updated openstack/openstack Project: openstack/keystone  f468f32d8f5fa3f01799e5ad3022eb8e065dc32a
Use idp_id and protocol_id in jsonhome

These are inconsistent with the rest of the federation routes. Note that
this doesn't change a public interface, it changes the path variables
names that exist within a jsonhome document which should only be
referenced within the same href_vars element and some internal variable
names.

Change-Id: I36d41e95d745bd98d36c3442bd4aa4faa91cab72
Closes-Bug: #1420125