Updated openstack/openstack Project: openstack/nova  2e0f4a9f15ac88333baa8f496ec0ce0148dd0ea3
Clean up legacy multi-version test constructs

This removes MyObj2 and MyObjDiffVers from the tests, and fixes up the
tests that still use them to use the more modern manifest-based approach,
which is much cleaner and doesn't require strangely subclassed objects
to mimic the multi-version environment during an upgrade.

It also removes a few tests that still do some basic validation on the
object registry, which is now fully contained (and tested) inside of
oslo.versionedobjects.

Change-Id: Ia523f3264b0ea54573e54c258efd7a07158ded36