Updated openstack/openstack Project: openstack/nova  78d07cce136467328eeabeeba90a8e914635ec85
libvirt: Log warning for wrong migration flag config options

The config options live_migration_flag and block_migration_flag
of the [libvirt] section set the flags for the hypervisor for the
live-migration operation. It is possible to set the flags in a way
that every live-migration will be a block-migration and vice versa.
This doesn't make sense and is a misconfiguration.

Because some customers could already have configured their systems in
this faulty way, this patch set logs a warning if it discovers this,
instead of blocking the nova-compute service to come up.

Closes-Bug: #1499041
Change-Id: I37f1e90c8a86a1dcf0fc1e699a018e0784c66e48