Updated openstack/openstack Project: openstack/ironic  9427a8c3fd5dfa6eb39be381e16b47e2be1bfd7f
Add clean up method for the DHCP factory

The DHCP provider base class has only update methods to notify the
provider which DHCP BOOT options to reserve when a node comes up. If anyone
wants to create his own DHCP provider implementation for Ironic, none of
the base class methods get triggered after a node is deleted, so the DHCP
reservations will be present even after the Ironic node is deleted.

This patch adds a 'clean_dhcp' method to the DHCPFactory class, which is used
to clean the DHCP BOOT options. This is called when an Ironic node is torn down.

Closes-Bug: #1498923
Change-Id: Iec768268323204b39fa80731e22a2511c7ac7763