Updated openstack/openstack Project: openstack/neutron  5fcfa65d974ba44924dcee89f7a7ccf195a1c606
Fix get_subnet_ids_on_router in dvr scheduler

Added a check to verify if we do have any elements in the list of
fixed_ips, before trying to retrieve the first element of the list, to
get the subnet id. There were no checks in the original code, so it
would crash.

Change-Id: If32db500aa3a0c299a5f19c33c05237e8e407e08
Closes-Bug: 1452458