Updated openstack/openstack Project: openstack-infra/irc-meetings  501a128bc2543f69dc9a383c3243a6ffd2105bb1
neutron_upgrades: added meeting_id and agenda_url

Also switched the channel to -3 since -2 is not an official channel with
logging. Note that to pass overlapping validations, the frequency is set
to bi-weekly since the channel is used by another project meeting
(zaqar) on even weeks. We'll need to follow up on the permanent solution
for the team (ideally, we would add another supported channel to reflect
huge expansion in the number of OpenStack projects due to the big tent).

Change-Id: Icb13409510212ec3353245c340f43e14fb87873a