Updated openstack/openstack Project: openstack/neutron  c1bbc0e45ec5001dff500c4c3f7d750c3e684c86
IPAM: fix 'enable-dhcp' with internal driver

Commit 0c1f96ad5a6606c1205bd50ea944c3a383892cde broke the ipam
internal driver. The base plugin should only call the ipam interface
and not invoke the IPallocation data model.

The code that validated that the subnet had one free IP address was
removed. This would not guarantee that a port would indeed be free,
that is, a parallel operation could have created a port on the subnet
during this check.

Closes-bug: #1514810

Change-Id: I4d7ecb985762a8eef39cb2f22f210b4899f1a22e