Updated openstack/openstack Project: openstack-infra/project-config  0f0b650a34d6a1ec582250f90cf88accefeac999
Add ansible-role-zuul project

In this patch, we are just adding ansible-role-zuul into the OpenStack
gerrit workflow. After talking with both openstack-infra and third
party CI teams it was decided to deal with which team this would be
part of at a later date.

Additionally, we have not created a launchpad project for this module
or pypi (since we won't release there).

Gerritbot was also not configured at this time, since we don't know
which team we are going to be part of.

Change-Id: I6424d66fb644c938e53091da11714e3f782f9a7c
Signed-off-by: Paul Belanger <pabelanger@redhat.com>