Updated openstack/openstack Project: openstack/tempest  5e4bbad4ceb43628056817b71d53b8df072787bc
Move identity wait until after password is updated

With Ied83448de8af1b0da9afdfe6ce9431438215bfe0 merging, we should be waiting
after the password is updated and before we reauthenticate. This change fixes
this to be correct. For some reason a transient failure didn't cause the tests
to fail.

Change-Id: I7286085325c3578958234176adb7588afdcac578
Closes-Bug: 1473567