Updated openstack/openstack Project: openstack/neutron  830c2e30fd716a34e1aa3f37a6e8f59fd8cc3824
Make sure we return unicode strings for process output

Process output is supposed to be represented with lines, so we should
put Python strings in the queue (not bytes). Just in case, we do it only
for Python 3 environment.

To fix that, we reuse code from utils.execute() linux/windows
implementations.

This fixes the TestAsyncProcess.test_async_process_respawns functional
test for Python 3 environment.

Related-Bug: #1515118
Change-Id: I9efec2290003add44909aab33a0026372a580016