Updated openstack/openstack Project: openstack/nova  f4e898b719a02af26938551059d8b14c76e7cd13
Fix resource tracker VCPU counting

The resource track is double counting the number of
vcpus used because it counts the virt driver's view
and the resource tracker's view. It should reset
the count to 0 before counting the resourcet racker's
view. this is done for other resources (memory, disk)
in _update_usage_from_instances().

This was not caught in unit tests because they always
provide a virt driver view with no resource used for
tests that assume instances are already running.

This patch adds the reset to 0 and modifies the
tests to provide correct counts of used resources
from the virt driver.

Change-Id: I21acfd7d5b5ed442cd1e9049c986a06bd2b59b81
Closes-Bug: #1519001