Updated openstack/openstack Project: openstack-infra/subunit2sql  7a16adcd2452e11161b634453f70c65ea140378d
Add reno release notes to subunit2sql

This commit adds support for using reno to generate the subunit2sql
release notes. This will make it much easier to keep up-to-date
and accurate release notes. Also doing it in tree will make the
notes reviewable which will help with that too.

Change-Id: I75502185a666945721dd84e4710e798962e612ee