Updated openstack/openstack Project: openstack/oslo.config  0ffdd4ebcdd036e2572cd3f228e50dc14678c341
Add 'mutable' param to Opt

re: https://etherpad.openstack.org/p/mitaka-cross-project-dynamic-config-services

oslo.config already offers a reload_config_files method. This patch adds
a 'mutable' flag to Opt, allowing services to mark which opts may be
reloaded. It also adds a flag to ConfigOpts, 'limit_mutation', to say
that when reload_config_files is called, only mutable opts should be
changed. If a new value for an immutable opt is detected, a warning is
issued.

Change-Id: I719874b58e64bc4f8535c85da43670779fbfd380