Updated openstack/openstack Project: openstack/heat  b92b349455e29df53e53a879edcd4de0eda7f545
Fix HTTP error codes due to invalid templates

This reverts commit 9a4d719ea1a6ca63f527b5c78308fd288058d42b and creates a
new exception type, InvalidGlobalResource, that is raised when a
non-existent template is referenced in the global environment and which
results in an HTTP 500 error.

This fixes the cases where we could return 404 or 500 errors as a result of
template validation failures, when we should return a 400 error.

Change-Id: I3c7b3182957448bb13514696cc2e12f5aaddd253
Closes-Bug: #1518458
Related-Bug: #1447194