Updated openstack/openstack Project: openstack-infra/shade  4c1418f7b415beac59a2d9d170a5ed420f81be06
Retry API calls if they get a Retryable failure

keystoneauth1 Session throws exceptions based on
RetriableConnectionFailure when the operation can be sanely retried.
Since we have commuication encapsulated, there is no reason to not
just retry.

We should maybe in the future allow for configuration of number of
times to retry - but let's start with one and go from there.

Change-Id: Iacc6ac3d7eecbccfa7602fefb238602cc8a66cc4