Updated openstack/openstack Project: openstack/horizon  ab7d5c3ec5317dea2495b958c9f82b2716ba4b5a
Top and Bottom Padding should be removed from navbar-brand

When using a theme other than 'default', a custom logo that is too
big won't be sized to fit inside of the header exactly perfectly.
The 'default' theme does this.  The style just needs to be moved
outside of 'default' theme to be global.

Change-Id: I241e49ad94cba24592209f5f4f23eedeb0e592a6
Closes-Bug: #1518123