Updated openstack/openstack Project: openstack/neutron  83134f899a02314d8530d0a45c26a41eaa3a6eeb
Check gateway ip when update subnet

If a IP address is None, the method netaddr.IPAddress will raise a
error as 'failed to detect a valid IP address from None'.
When it updates subnet, if gateway_ip is None, it should not call the
method validate_gw_out_of_pools.

Change-Id: Ibb007fe7e6402c174129074288b8af891182cd68
Closes-bug: #1511925