Updated openstack/openstack Project: openstack/nova  8970fc9e4f7942a2f6bfb1fe653dd5cb8061e71e
docs: describe migration and other movement concepts

Migration (including cold and live migration), resize and shelve
are three concepts at the API level that support three distinct
sets of use cases. These all involve movement of servers and
so provide opportunity for operational efficiency.

This patch adds a section to the server concepts that outlines
the purpose behind the actions and the operational efficiencies
that can be gained by them.

blueprint complete-todo-in-api-concept-doc

Change-Id: I6b9c6c179c1bdd9400abae7deb1918b370ddaefd