Updated openstack/openstack Project: openstack/cinder  993aaa4c6ab3c5c1b795ad3f9a08e13b7fac0ed1
Fix metadata retrieval in GPFS driver

GPFS driver was using 'volume_metadata' to get the
volume metadata. With versionedobjects, it is to be
retrieved using 'metadata' from volume.
This patch fixes the issue. It first tries for 'volume_metadata'
in the volume object, if not present, it will try for 'metadata'
to get the metadata.

Change-Id: I97649874ab54187a469193c8cae773409e9913e6
Closes-bug: 1517404