Updated openstack/openstack Project: openstack/python-zaqarclient  8ca747746c1ea20cbc5288b00d030359aa2fd7c4
Add support for subscription create v2

This patch adds the support for subscription create, which includes
the CLI support and the lib support. Besides, to make the functional
test work for subscription create, this patch also adds the lib
support for subscription get, delete, update and list to make things
easier.

Change-Id: Ie3ac27b731f41ea6023976aab3c1afbc283d659d