Updated openstack/openstack Project: openstack/cinder  2c3b6e9912b6dc60b78585effaff6219377e5d8d
Declare multiattach is True in RBD driver

An RBD volume can be attached to multi instances, however,
this feature is not declared in RBD driver. As a result,
if a user create a multi-attached volume using RBD as
backend, they will be failed.

In this commit, we declared multiattach is True in RBD
driver.

Change-Id: I70c2c0528d9c00629f9cf34d79262a5bf4750e34
Closes-Bug: #1519227