Updated openstack/openstack Project: openstack-infra/system-config  00a0d57847e526dcbcee88880c812e0885b41eab
Add puppet apply settings to puppet.conf

We need to exist for a period of time with both agent and apply being
operational so that we can test things appropriately.

This moves agent specific settings to the [agent] section and adds a
[user] section which is used to control puppet apply.

As part of this, we need to add a production environment to all of our
nodes. Doing this in this way will also cover the current puppetmaster,
since puppetmaster is a puppet client.

Change-Id: I550c474d1c51c5795f745630fb91ee8cc1a55e36