Updated openstack/openstack Project: openstack/nova  b0954683fbd0861a782e6594f686768785567d28
Unify on _schedule_instances

There's already a helper method for unshelve_instance. It turns out
that the other two places in this file which call select_destinations
can use the same helper. Doing this removes some code redundancy.

I haven't pushed build_request_spec in as well because for bp
check-destination-on-migrations we won't always be constructing a
brand new request spec.

Change-Id: I19bad96d997f61958e32f6254bc4a428395bc7ec