Updated openstack/openstack Project: openstack-infra/release-tools  de4a9318212b0e8ce1329eafebd6f6d60e6e0e2b
fix error handling and bug reporting in release script

After the tag is pushed, we don't want to die with an error if we can't
update the bug reports, so turn off the -e shell flag.

We have to check for the previous tag before we add the new one, because
the logic for finding the last tag assumes the new one is not there yet.

Change-Id: If2ac25b8bd96e4e0c5b90016a3c40d43733ab849