Updated openstack/openstack Project: openstack-infra/subunit2sql  ac54f82a807cb521e6638f7880a43bd92f72473b
Fail early if no data is found in the database

For the run_time graph subcommand if invalid user input was recieved
and there was no data in the database we previously would continue
running until something in the data processing would raise an
exception because there was no data present. This error was often
cryptic and hard to debug. This commit fixes this by checking if
data is returned from the db api calls and if it's not then we
print an error message explaining what data was not found and
exit with a non-zero code.

Change-Id: I460679f90a46244ee0c0818bf7b3deee4cce280c