Updated openstack/openstack Project: openstack-infra/project-config  f36c02efac29d146933f52a32a93deb1f0c70d91
Set a temporary cap on subunit2sql before 1.0.0 release

This commit adds a temporary version cap on subunit2sql to be < 1.0.0.
The 1.0.0 release includes a very large database migration which will
be slow to execute. The python DB api from >=v1.0.0 will not work with
a database that doesn't have the updated schema. So while the migration
is running let's cap the version we install to prevent everything from
breaking while the migration is running. (which might take days)

Change-Id: Iab89beb5c7aba8b744a62f5063e513b72cab0ec2