Updated openstack/openstack Project: openstack-infra/puppet-jeepyb  4b3de83a24ab21e3be305e11d7e0d9877e1f33e4
Remove mysql::python class inclusion

When the dependent patch is merged, Jeepyb will depend on PyMySql and
install it from pip as part of its requirements. Therefore it will no
longer be necessary to include the mysql::python class, which installs
the python-mysqldb package from Ubuntu packages. Removing this class
inclusion will make transitioning to an up-to-date version of the mysql
puppet module easier, as the current version moves the mysql::python
class to mysql::bindings::python.

Change-Id: I7dec1c6f588cb3e1d0eb3cb2eb6398df45ebdc22
Depends-On: I3afeaebeb8416388d5866fc3994357935079c97d