Updated openstack/openstack Project: openstack/horizon  705995f6088a04a75f0014c1a5f022cf4e01cae0
Fix flavor quota charts in NG Launch Instance

The flavor quota charts aren't rendering correctly
when the instance count is changed in the Angular
Launch Instance modal. The <svg> element was not
available to add slices as 'chartData.showChart'
is originally undefined.

This patches fixes the the issue by using ng-show
instead of ng-if.

Change-Id: I7b24da7942c3d02a55adc732d70a26aeb5b623ec
Closes-Bug: #1518030