Updated openstack/openstack Project: openstack/keystone  212c905b3c36883d3c03d8fa54f814c58569762e
Correct docstring warnings

There were warnings printed when docs were generated:

 docstring of keystone.assignment.core.Manager.create_grant:8:
 ERROR: Unexpected indentation.

These are because the docstrings were not valid rst. The docstrings
are corrected.

In the case of create_grant, the problem is that the docstring of
the wrapper was used, so @functools.wraps is added so that the
docstring of the function is used rather than the wrapper. Also,
the wrapper's docstring was incorrectly formatted.

Also, corrected other issues with docstrings.

Change-Id: Ide6a85365f71330bf4caa24babc3bfe56174fc0f