Updated openstack/openstack Project: openstack/ironic  7febdccd94c947f70faa4c197da73c76c93d48eb
refine the ironic installation guide

From the Juno version, we have the OpenStackClient to identify
the component. So we can substitute OpenStackClient for keystone
to identify the ironic.
I checked the OpenStackClient commands guide:
https://wiki.openstack.org/wiki/OpenStackClient/Commands#endpoint
The OpenStackClient API version is:API v2.0
The OpenStackClient command is:
os endpoint create
    --publicurl <public-url>
    [--adminurl <admin-url>]
    [--internalurl <internal-url>]
    [--region <endpoint-region>]
    <service>
And the keystone command is:
keystone endpoint-create
    [--region <endpoint-region>]
    [--service_id <service-id>]
    [--publicurl <public-url>]
    [--adminurl <admin-url>]
    [--internalurl <internal-url>]
Change-Id: I51ced0aeb04aed486cbe644305a9c19233e1d398