Updated openstack/openstack Project: openstack-infra/irc-meetings  7a10e641c3aae5e69a0ef5f54a13329b232a9c48
Add check_channels.py and testenv to run it

Create a new tool called check_channels.py that ensures that the irc
channel is one which is capable of running a fully functional MeetBot
*or* one that is an 'office hours' type event that doesn't use MeetBot.

Also create a tox textenv to run it.  Once this is clean we can add a
gate job to run these checks and vote.

Change-Id: I57a725b55fb611214d0f0eec08179aae01c1e49a