Updated openstack/openstack Project: openstack-infra/project-config  47a60683d4974573208258097007c19baabf760b
networking-bgpvpn: remove check-requirements jobs

As a follow up to I21c8aadd82249882243bf8e23b37e23fcb62f507,
and after an IRC discussion with the infra team [1], there seems
to be an agreement that simply removing the check-requirements
job from networking-bgpvpn is the best *temporarily* solution,
until a decision is made on how a project like ours should work
in branches.

We plan to add automated requirements checks when it will be clear
how to do it in our different branches.

This change is dependent on Ibc3fb11598298938bc72be2ce8fc12d5e231b6bc
to guarantee that we won't propagate bogus requirements in global
requirements.

Also related: email thread starting at
http://lists.openstack.org/pipermail/openstack-dev/2015-November/079773.html

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-infra/%23openstack-infra.2015-11-24.log.html#t2015-11-24T16:03:50

Change-Id: I60df289591a11b53de5f15835137892e32873b15
Depends-On: Ibc3fb11598298938bc72be2ce8fc12d5e231b6bc