Updated openstack/openstack Project: openstack/manila  14865090e2e0477198f2bd933cd2b6ef1fac1f22
glusterfs/volume layout: fix incorrect usage of export_location

Actually, all uses of export_location are incorrect -- from the
layout code's point of view, export_location is an arbitrary
opaque value, obtained from self.driver._setup_via_manager with
which the only legit action is to return it from create_share*.

That we use export_location as dict key in ensure_share is just
a pre-layout relict that survived by the virtue of remaining
unnoticed. Now the referred bug forced it out of the dark.

Change-Id: I965dae99486002f00145daff0cd2a848777b5b81
Partial-Bug: #1501670