Updated openstack/openstack Project: openstack/cinder  cdc645e2f9da5dccd7d05eb157029e2542ac18db
Fix ScaleIO driver provisioning key

In the ScaleIO Cinder volume driver, the parameter 'sio:provisioning'
is wrong. It should be 'sio:provisioning_type' instead. As a result of
this error, the volume is always provisioned as 'thick' even if user
wants to provision a 'thin' volume.

This patch corrects this parameter.

Closes-Bug: #1519611
Change-Id: I7c3b0256af3c30c2f375c0f9373646d3f58cd0d0