Updated openstack/openstack Project: openstack-infra/shade  8cf3c81b2594bfb9cccecbab1c845dcd44edee63
Fix resource leak in test_compute

Depending on the time between the create_server() and
delete_server() calls, it's possible the delete does not
actually find the server we want to delete, so we leave
instances hanging around after our tests. Fix this by waiting
for the servers with wait=True to create_server().

Change-Id: I908fdb94dcf5e410597a7cd3ecc6c962f2bfbce3