Updated openstack/openstack Project: openstack/nova  3c67982f7c8c95da335310348fda41e2cf996f3a
SR-IOV: Improve the vnic type check in the neutron api

The Neutronv2 API checks the vnic_type to see whether
or not it should include a PCI requests for SR-IOV ports.
The check is "if the vnic_type is not NORMAL" which is fine
for now that we have only NORMAL, DIRECT and MACVTAP, but in the future more vnic types
can be added. This patch update the check to be what the vnic types
are expecting to be.

Change-Id: I7b8ba1acfbafc3d02ab50afadf4627fcce052484