Updated openstack/openstack Project: openstack/ironic  ce3878176e20d14681d11b53f4fc5cf1cb63e956
Add switch to enable/disable streaming raw images for IPA

This patch is adding a configuration option called "stream_raw_images"
under the "agent" group. By default this option is set to True, meaning
that the IPA ramdisk will be instructed to stream raw images directly
onto the disk instead of copying it to a tmpfs partition prior of
writing it to the disk. By streaming raw images IPA will consume less
memory to run and the deployment will be faster (unless the disk being
deployed is really slow).

Depends-On: Iddf67907bc9b54bbd3065a97064cb5a3602cfe18
Related-Bug: #1505685
Change-Id: I7e7ed7e7aeebbf7c936605a72aa974db2123a43e