Updated openstack/openstack Project: openstack-infra/nodepool  5bcccf76050252a1ce4aade48989c6d85f7ca48d
Suppress NoValidConnectionsError from paramiko

New versions of paramiko wrap exceptions from multiple connection
attempts for multiple address families into one
NoValidConnectionsError exception.  It is a subclass of socket.error
but with an errno set to None.  Just check for that and ignore it
to supress log entries on perfectly normal connection failures.

Change-Id: If64ab66dcc6db7c1886fb72f36078f7f819d6506